triplanetary/requirements.txt
boto3>=1.28.0
python-dotenv
